:recycle: rename files
