+++ elements/reigns/src/screens/EndedScreen/EndedScreen.tsx
@@ -1,10 +1,7 @@
-import React from "react";
-import { Header } from "../../Header";
 import { Card, GameDefinition } from "../../features/game/types";
-import { useTextFit } from "../../useTextFit";
 import { getEndMessage } from "./utils";
-import { StartedScreen } from "../StartedScreen";
-import { useVotes } from "../ConnectedStartedScreen";
+import { VotingScreen } from "../VotingScreen";
+import { useVotes } from "../../useVotes";
 import { GamePhase } from "../../constants";
 
 export function EndedScreen({
@@ -12,20 +9,18 @@ export function EndedScreen({
   currentStats,
   isGameWon,
   round,
-  doRestartGame,
 }: {
   gameDefinition: GameDefinition;
   currentStats: number[];
   isGameWon: boolean;
   round: number;
   isHost: string | boolean | undefined;
-  doRestartGame: () => void;
 }) {
   const endMessage = getEndMessage(gameDefinition, currentStats, isGameWon);
   const votes = useVotes();
 
   return (
-    <StartedScreen
+    <VotingScreen
       gameDefinition={gameDefinition}
       {...votes}
       selectedCard={
@@ -36,7 +31,6 @@ export function EndedScreen({
       }
       currentStats={currentStats}
       round={round}
-      doRestartGame={doRestartGame}
       phase={GamePhase.ENDED}
     />
   );
